+++ data/mypage/src/main/java/org/sopt/official/data/mypage/model/response/UserResponse.kt
@@ -26,7 +26,6 @@ package org.sopt.official.data.mypage.model.response
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import org.sopt.officiail.domain.mypage.model.SoptampUser
 
 @Serializable
 data class UserResponse(
@@ -38,5 +37,5 @@ data class UserResponse(
     val profileMessage: String? = null,
 ) {
     fun toDomain() =
-        org.sopt.officiail.domain.mypage.model.SoptampUser(nickname, points, profileMessage ?: "")
+        org.sopt.official.domain.mypage.model.SoptampUser(nickname, points, profileMessage ?: "")
 }
