+++ feature/mypage/src/main/java/org/sopt/official/feature/mypage/soptamp/nickName/ChangeNickNameViewModel.kt
@@ -33,7 +33,7 @@ import io.reactivex.rxjava3.subjects.PublishSubject
 import kotlinx.coroutines.launch
 import org.sopt.official.common.util.rx.subscribeBy
 import org.sopt.official.common.util.rx.subscribeOnIo
-import org.sopt.officiail.domain.mypage.repository.UserRepository
+import org.sopt.official.domain.mypage.repository.UserRepository
 import timber.log.Timber
 import javax.inject.Inject
