+++ feature/mypage/src/main/java/org/sopt/official/feature/mypage/soptamp/sentence/AdjustSentenceViewModel.kt
@@ -29,7 +29,7 @@ import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import io.reactivex.rxjava3.subjects.PublishSubject
 import kotlinx.coroutines.launch
-import org.sopt.officiail.domain.mypage.repository.UserRepository
+import org.sopt.official.domain.mypage.repository.UserRepository
 import timber.log.Timber
 import javax.inject.Inject
