+++ feature/soptamp/src/main/java/org/sopt/official/stamp/feature/setting/SettingScreenViewModel.kt
@@ -31,7 +31,7 @@ import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.launch
 import org.sopt.official.domain.soptamp.repository.StampRepository
-import org.sopt.officiail.domain.mypage.repository.UserRepository
+import org.sopt.official.domain.mypage.repository.UserRepository
 import timber.log.Timber
 import javax.inject.Inject
